refactor: add Storybook util to define argType as icon dropdown
